requirements.txt
alabaster==0.7.6
Babel==2.1.1
docutils==0.12
Jinja2==2.8
MarkupSafe==0.23
Pygments==2.0.2
python-dateutil==2.4.2
pytz==2015.6
six==1.10.0
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9
